requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
idna==2.8
mypy==0.761
mypy-extensions==0.4.3
pathspec==0.7.0
psutil==5.6.7
python-dotenv==0.10.3
PyYAML==5.3
regex==2020.2.20
requests==2.22.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
